primary_parser/requirements.txt
vk_api
requests
tqdm
pandas
torch
transformers
Pillow
psycopg2-binary
python-dotenv

